+++ RelationalAI.Examples/GetTransactionMetadata.cs
@@ -22,7 +22,7 @@ namespace RelationalAI.Examples
                     Description = "Profile name from .rai/config to connect to RAI Cloud."
                 }
             };
-            cmd.Description = "Get an asynchronous transaction metadata.";
+            cmd.Description = "Get an asynchronous transaction protobuf metadata.";
             cmd.Handler = CommandHandler.Create<string, string>(Run);
             return cmd;
         }
@@ -32,11 +32,9 @@ namespace RelationalAI.Examples
             var config = Config.Read("", profile);
             var context = new Client.Context(config);
             var client = new Client(context);
-            var metadataList = await client.GetTransactionMetadataAsync(id);
-            foreach (var metadata in metadataList)
-            {
-                Console.WriteLine(metadata);
-            }
+
+            var metadata = await client.GetTransactionMetadataAsync(id);
+            Console.WriteLine(metadata);
         }
     }
 }