+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.6.0-alpha
+* Deprecated metadata json format.
+* Removed `TransactionAsyncMetadataResponse` model.
+* Added support to metadata protobuf format.
+* `GetTransactionMetadata` returns protobuf metadata.
+
 ## v0.5.0-alpha
 * In this update we ensure the SDK follows the standard async/await approach for the methods. The following methods have been renamed to include the "Async" postfix, made async and now return a `Task<T>` to the user:
   - `CreateDatabase`, `GetDatabase`, `ListDatabases`, `DeleteDatabase`, `CloneDatabase`.